+++ packages/stream_chat_flutter/test/src/attachment/giphy_attachment_test.dart
@@ -38,6 +38,7 @@ void main() {
                     extraData: const {
                       'mime_type': 'gif',
                     },
+                    uploadState: const UploadState.success(),
                   ),
                 ),
               ),