+++ packages/stream_chat_flutter/lib/src/message_widget/message_widget_content.dart
@@ -259,9 +259,7 @@ class MessageWidgetContent extends StatelessWidget {
                     reverse ? CrossAxisAlignment.end : CrossAxisAlignment.start,
                 mainAxisSize: MainAxisSize.min,
                 children: [
-                  if (message.pinned &&
-                      message.pinnedBy != null &&
-                      showPinHighlight)
+                  if (isPinned && message.pinnedBy != null && showPinHighlight)
                     PinnedMessage(
                       pinnedBy: message.pinnedBy!,
                       currentUser: streamChat.currentUser!,