+++ packages/stream_chat_flutter/lib/src/theme/stream_chat_theme.dart
@@ -308,6 +308,7 @@ class StreamChatThemeData {
             colorTheme.disabled,
           ],
         ),
+        useSystemAttachmentPicker: false,
       ),
       galleryHeaderTheme: StreamGalleryHeaderThemeData(
         closeButtonColor: colorTheme.textHighEmphasis,