+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Upcoming
+
+🐞 Fixed
+
+- [[#2118]](https://github.com/GetStream/stream-chat-flutter/issues/2118) Fixed an issue where message input buttons could not be customized.
+
 ## 9.4.0
 
 🔄 Changed