+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Upcoming
+## 9.5.0
 
 🐞 Fixed